requirements.txt
astroid==1.6.3
autopep8==1.3.5
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
django-js-asset==1.1.0
django-mptt==0.9.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.4.0
pylint==1.8.4
pytz==2018.4
six==1.11.0
wrapt==1.10.11
